Added jacoco profile
The configuration is the following:
- Added the plugin and bind it by default to initialize (prepare-agent)
and verify (report).
- Excluded model classes from coverage
- Created a profile named jacoco that activates the profile.

Running mvn clean verify -Pjacoco will execute tests and generates the
report in target/site/jacoco.

Note that the build will not fail if the code coverage is not at a
certain level because it is not configured to do so. It will be done by
coveralls in future commits.
Signed-off-by: Christian Mancini <mancinichristian00@gmail.com>